+++ CHANGES.md
@@ -1,3 +1,6 @@
+0.11.0
+======
+
 * Add Graphite.Report module to generate reports for easy printing
 
 0.10.0